[WEB-72] ✨ feat: SearchPageTemplate 분리
