+++ src/app/search/components/SearchResultCount.tsx
